Split StepSummaryWriter into partial files
